+++ CHANGELOG.md
@@ -199,3 +199,4 @@
 - 2023-12-06: Shutdown contacts, daemon, follow up, descend, +4
 - 2023-12-07: Add follow up, configuration, contacts, lib, +5 db
 - 2023-12-07: Features NOTICE, campaign, ready pool, geo, +2
+- 2023-12-08: Bench tz country, main, campaigns, sender, +6